+++ Engine/Rocket.h
@@ -21,7 +21,7 @@ public:
 	RectF getHitbox() const;
 	const VecF2& getPosConst() const;
 	int getType() const;
-	int getAttack() const;
+	float getAttack() const;
 private:
 	VecF2 pos;
 	VecF2 vel;
